Fix ChunkPrivacyMixin
It turns out you can't really mixin classes that are indeterminately loaded sometime after initial load. Well, it *might* be possible, but it's probably a huuuuge faff. So the fix here is to mixin the initial handler from the mod-class which then relays the chunk data to the remote-loaded class.